requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
marshmallow==3.21.3
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
